requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
gunicorn==20.1.0
pandas==1.4.3
